fix: update `/knowledgebase` to `/docs`
